pyproject.toml
[tool.poetry]
name = "hackdefyproject"
version = "0.1.0"
description = ""
authors = ["Replit user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask-socketio = "^5.0.1"
python-dotenv = "^0.16.0"
firebase-admin = "^4.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
